requirements.txt
numpy
pillow
opensimplex
argparse
